requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.1.0
python-decouple==3.8
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
razorpay==1.4.1
celery==5.3.4
django-celery-beat==2.5.0
boto3==1.29.7
django-storages==1.14.2
requests==2.31.0
twilio==9.7.0